IMplement an Echo Manager
