+++ src/js/component/index.js
@@ -10,6 +10,7 @@ import PrefetchManager from './PrefetchManager'
 import MethodAction from '@/action/method'
 import ModelAction from '@/action/model'
 import MessageBus from '../MessageBus'
+import EchoManager from './EchoManager'
 
 export default class Component {
     constructor(el, connection) {
@@ -30,12 +31,13 @@ export default class Component {
         this.modelTimeout = null
         this.tearDownCallbacks = []
         this.prefetchManager = new PrefetchManager(this)
+        this.echoManager = new EchoManager(this)
 
         store.callHook('componentInitialized', this)
 
         this.initialize()
 
-        this.registerEchoListeners()
+        this.echoManager.registerListeners()
 
         if (this.redirectTo) {
             this.redirect(this.redirectTo)
@@ -341,47 +343,6 @@ export default class Component {
 
             callback(el)
         })
-    }
-
-    registerEchoListeners() {
-        if (Array.isArray(this.events)) {
-            this.events.forEach(event => {
-                if (event.startsWith('echo')) {
-                    if (typeof Echo === 'undefined') {
-                        console.warn('Laravel Echo cannot be found')
-                        return
-                    }
-
-                    let event_parts = event.split(/(echo:|echo-)|:|,/)
-
-                    if (event_parts[1] == 'echo:') {
-                        event_parts.splice(2,0,'channel',undefined)
-                    }
-
-                    if (event_parts[2] == 'notification') {
-                        event_parts.push(undefined, undefined)
-                    }
-
-                    let [s1, signature, channel_type, s2, channel, s3, event_name] = event_parts
-
-                    if (['channel','private'].includes(channel_type)) {
-                        Echo[channel_type](channel).listen(event_name, (e) => {
-                            store.emit(event, e)
-                        })
-                    } else if (channel_type == 'presence') {
-                        Echo.join(channel)[event_name]((e) => {
-                            store.emit(event, e)
-                        })
-                    } else if (channel_type == 'notification') {
-                        Echo.private(channel).notification((notification) => {
-                            store.emit(event, notification)
-                        })
-                    } else{
-                        console.warn('Echo channel type not yet supported')
-                    }
-                }
-            })
-        }
     }
 
     modelSyncDebounce(callback, time) {
